requirements.txt
# --- Core CLI & UI ---
click
colorama
rich

# --- Scraping Engine (Fast & Fallback) ---
requests
beautifulsoup4
selenium
webdriver-manager
aiohttp
curl_cffi

# --- Data & Output Management ---
pandas
openpyxl

# --- Google Cloud / Sheets ---
gspread
google-auth-oauthlib
google-api-python-client
google-genai
